bamss-tests/requirements.txt
locustio
pyzmq
msgpack-python
